requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.23
celery==3.1.23
click==6.6
Django==1.9.7
django-celery==3.1.17
django-redis==4.4.3
djangorestframework==3.3.3
kombu==3.0.35
ordereddict==1.1
Pillow==3.2.0
psycopg2==2.6.1
python-dateutil==2.5.3
python-dotenv==0.5.1
pytz==2016.4
redis==2.10.5
requests==2.10.0
six==1.10.0
wheel==0.24.0

